added power analysis to report power in manuscript
due to the very small sample size it would be useful to add the potential power of the study to a manuscript. How likely is it that 8 or 11 trips would BY CHANCE have recorded zero bycatch? I have only done this for birds here, but you can do it for other fish as well.